refactor: Cut out functions related to the modal to a separate file.
